Updates to request display to add funcitonality

+++ ethereum/contracts/Campaign.sol
@@ -88,7 +88,7 @@ contract Campaign {
         );
     }
 
-    function getRequestsCount() public view returns (uint) {
+    function getRequestsCount() public view returns (uint256) {
       return requests.length;
     }
 }
